fix test suite: load .env.test, fix XML POST/PUT, fix Product pagination
- Add tests/bootstrap.php to load .env.test via phpdotenv with
  PutenvAdapter so getenv() works in test setup
- Fix TypeError in HTTPXMLResource post()/put() by converting data
  arrays to XML strings via new arrayToXml() method
- Fix Product::processResponse() to throw ApiException instead of
  RuntimeException for unparseable XML responses
- Default startRow/pageRows on Product collection requests to prevent
  AP21 database errors when pagination params are omitted
- Disable countEnabled on Product (uses TotalRows XML attribute instead)
- Fix usleep() float-to-int deprecation in Ap21SDK::checkApiCallLimit()
- Refactor TestSimpleResource to use markTestSkipped() and remove
  testGet dependency on testPost

+++ lib/Ap21SDK.php
@@ -190,7 +190,7 @@ class Ap21SDK
 
         if ($timeToWait) {
             //convert time to microseconds
-            $microSecondsToWait = $timeToWait * 1000000;
+            $microSecondsToWait = (int) ($timeToWait * 1000000);
             //Wait to maintain the API call difference of .5 seconds
             usleep($microSecondsToWait);
         }
